requirements.txt
apache-airflow==2.9.1
apache-airflow-providers-google==10.17.0
apache-airflow-providers-slack==8.7.1
apache-airflow-providers-http==4.10.0

# Data Warehousing
google-cloud-bigquery==3.20.0
google-cloud-bigquery-storage==2.25.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
db-dtypes==1.2.0

# ETL / Data Processing
pandas==2.2.2
pyarrow==16.1.0
numpy==1.26.4
python-dateutil==2.9.0

# API Clients — Marketing Platforms
google-ads==24.1.0
facebook-business==20.0.2
google-analytics-data==0.18.9

# CRM
hubspot-api-client==8.2.1
simple-salesforce==1.12.5

# E-commerce
ShopifyAPI==12.6.0

# Data Transformation & dbt
dbt-core==1.8.2
dbt-bigquery==1.8.2

# Data Quality
great-expectations==0.18.15
pandas-gbq==0.22.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
requests==2.31.0
tenacity==8.3.0
structlog==24.1.0
click==8.1.7

# Scheduling & Monitoring
croniter==2.0.5

# Testing
pytest==8.2.0
pytest-mock==3.14.0
pytest-cov==5.0.0
responses==0.25.0
freezegun==1.5.0
